Prevent ShaderBuildMessage from being exported twice to QML
We need to export the type as a value type, and also export its enums
in a namespace. This resulted in duplicated entries in the qmltypes
file, which confuses Qt Design Studio.

Fix it by introducing a derived class, and using it for exporting the
enums.

Fixes: QTBUG-115362
Change-Id: I7a052592961fcbcecb8b48757a9babaa79d81f77
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 17d9ce3f2abc318366a4fa428c16a0baa6f7812a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>