Fix incorrect usage of QSSGShaderFeatures in QSSGShaderMapKey
Since 052021833 the QSSGShaderFeatures have been optimized to be stored
as unique 4byte key, meaning we avoided expensive hashing. But the
QSSGShaderMapKey was not using it as such, which meant we were not
getting any benefits of the optimization and since the address of the
feature set was used, there was close to a zero percent chance to
reusing shaders with the same properties and features.

Pick-to: 6.6 6.5
Change-Id: Ief5d0a6efeb1dc0ca3a38b6132a9777d38facd19
Reviewed-by: Andy Nichols <andy.nichols@qt.io>