Improve ProceduralSky 3rd party attribution docs
This patch makes it more clear that the ProceduralSky component is
also related to the GodotEngine (similar to the Fog).

ProceduralSky isn't a project by itself, but we are attributing
that the code we're using for our ProceduralSky came from the
GodotEngine. This is not something we need to update going forward,
the attribution just needed some clarification.

Task-number: QTBUG-116247
Change-Id: I23daffdffe3be05f33175b793a4ad278f34aa432
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 0eadaac4f19f7a2357376ab71113ecad792aa56c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>