Improve the draw call data key
Removed the selector entry in the key, as we don't want to use a fixed
value for it. We have also been using the QSSPassKey
(address of the sub-pass) as an entry in draw call data key for quite some
time already.
The usage of the cube map faces where also improved, with a defined
value for "none"/"unused" and some better inline explanations of how we
use it in the dcd key.

Change-Id: I679d778edc402c3e448caef695aea90bf369b745
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 9d5543cfd24b49bc721e8773e90a6c3bb3f623aa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>