Add a way to override the backing texture size with the Offscreen mode
Consistent with QQuickRhiItem and QRhiWidget, and also exposed in the
DebugView.

[ChangeLog] Added View3D properties to override the backing texture's
size when using the default Offscreen render mode. This allows the
application to specify its own render target resolution (for example
because the developer knows that the content is fine with a lower
resolution instead of following the windowing system dictated
geometry and high DPI scale factor). For easy experimenting, the
properties are also exposed in the DebugView item.

Change-Id: I5f19b8a50af54ff692ecd40b336a7097f7e7e357
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>