Add extra sanity check to the attachWatcher function
This avoids triggering undefined behavior when using a scene context
that is already destroyed (QObject dtor). This was discovered when
testing QtGraphs, where some of the nodes are created manually and
then the scene root is used as a parent for new nodes.
In the case of the camera node that becomes problematic, as the root
node is also used as the scene context when setting up the connection
for watching the object, meaning the call-context (the view), is still
alive but the scene-context (the scene root object) has decayed into a
QObject.

Pick-to: 6.6 6.5
Change-Id: I81ac98e48fc6a716b3ca5ef098936c634799d366
Reviewed-by: Andy Nichols <andy.nichols@qt.io>