wasd: Accept the same set of buttons for drag and tap handlers
Setting acceptedButtons: Qt.RightButton on a WasdController continues to
eat the left button presses right now (due to the TapHandler, whose sole
purpose is to force the focus to ensure keyboard events). This is not
ideal when the whole point of moving view movement to the right button
is to free up the left button for other purposes.

Change-Id: I0bca38d7466f631d5a1d92ab8f40178530a2e89b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit f889716014d1f36a14220dc22fc128aea14299a6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>