Change constexpr float kAreaEpsilon to const
Because on VxWorks `FLT_EPSILON` is a macro pointing to union, which
field value can't be assigned to constexpr variable, change the
kAreaEpsilon to const. This is safe because it is internal of
xatlas.cpp.

Add patch describing this change between source of this 3rdparty and Qt
version.

Task-number: QTBUG-115777
Change-Id: I644186e9edd37b2e5493ba8e5be0a5c18754661e
Reviewed-by: Kimmo Ollila <kimmo.ollila@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>