Define RAPIDJSON_HAS_CXX11_RVALUE_REFS=1 for VxWorks
RapidJSON library tries to detect automatically if C++11 r-value
references are available, to which it tries to detect if standard
library is one of MSVC, libc++ or libstdc++ with specific version.
Because VxWorks toolchain uses custom standard library, this automatic
detection fails, and leaves RAPIDJSON_HAS_CXX11_RVALUE_REFS equal to 0.
This in turn causes compilation error in glTF2Asset.inl of assimp,
which relies on move constructor existence of
`rapidjson::GenericDocument`, but only private copy constructor is
available.

To fix that, define RAPIDJSON_HAS_CXX11_RVALUE_REFS=1 for VxWorks when
compiling assimp plugin.

Task-number: QTBUG-115777
Change-Id: Iafc60ca1c01eef60e71056da1c97c0c8f691bc78
Reviewed-by: Kimmo Ollila <kimmo.ollila@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>