Update GLSL versions for lightmapuvraster shaders
This is to fix lightmap baking when using OpenGL backend. The default
versions that qt_internal_add_shader set won't work (120 and 150).
Version 150 is used when a core profile context is requested and this
version does not have the location layout qualifier that multiple render
targets require. So we stick to 120, but change 150 to 330.

Pick-to: 6.5
Change-Id: I9b06b33d6fb1b8922da19fc94a07bc3bf67803e2
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 677d772827ec24a8ced232467fe0aea381135781)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>