Add rhi rendering extension manual test
Have a green cube moving around in the scene. Get the view projection
matrix from the camera. Apply the QRhi-provided correct matrix
correctly. Show a simple custom property and its synchronization with
the extension backend.

Demonstrate a few concepts, such as how to prepare for a dynamically
changing MSAA setting, or how to prepare if the View3D's render target
changes, possibly invalidating the custom renderer's
existing QRhiGraphicsPipeline objects.

The render mode is dynamically changeable, to allow seeing the effect
of switching between Underlay and Overlay. (visible when the material
for the View3D scene objects is semi-transparent)

Task-number: QTBUG-116570
Change-Id: I5f139da5d83edfa161a604da4d93b96d83300d58
Reviewed-by: Christian Strømme <christian.stromme@qt.io>