Make sure that View3D cleans up resources as it is destroyed
Previously the buffer manager would retain references to all of its
created buffers at destruction until the QQuickWindow it was associated
with was destroyed. This is undesirable because if you create and
destroy View3D items at runtime anything that was in use at destruction
would just hang around in GPU memory and could only be deleted
explicitly. This patch makes sure to clear all references in the buffer
manager and clean up anything that is not also referenced by another
View3D owned by the QQuickWindow.

Fixes: QTBUG-116136
Pick-to: 6.5 6.6
Change-Id: I73946693e9937f677d5f6639ae5564271b8be2f5
Reviewed-by: Christian Strømme <christian.stromme@qt.io>