Do not crash when an extension provides no spatial node
Rather, ignore it.

This is half of the story. The other half would be to recognize
that an extension that previously provided no spatial node now
has one, and trigger the rebuilding of the extension list.
That can currently be done only by manually calling the
(undocumented) rebuildExtensionList() on the View3D.

Change-Id: I24594db6359dde514d5b955a810f12e4535549ea
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>