Fix View3D implicit input handling for Texture with sourceItem
The issue with the current logic seems to boil down to a timing issue,
where m_qsgTexture is not yet set in updateDirtyResource, but is
available in cleanupNodes, making the decrement happen, but not the
increment. This results in inputHandlingEnabled going into the
negatives. Checking on m_qsgTexture is also unnecessary, as we already
have an explicit type to check for (Image2D). This change simplifies the
logic greatly to be identical to the handling of Item2D in
updateDirtySpatialNode.

Fixes: QTBUG-117023
Pick-to: 6.5
Change-Id: I96291bf53e48cbc525b08630e5901441845897ca
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
(cherry picked from commit 15abed925b55a0838fe66cecbe9826fc53e6ae4d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>