Expose the render context interface
The QSSGRenderContextInterface is the main hub for all the internal
sub-components. Note that while the RCI is exposed, it doesn't mean all
the sub-components are exposed as well. Depending on the component some
might be exposes with a semi-public API while others still will require
the use of private headers.

Task-number: QTBUG-116570
Change-Id: Iad0d0069fbdfd8b66af6f3d020659395989b49dc
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>