Further improvements to the BVH code
Avoid the chance of invalid handles returning a null pointer by
restricting how node handles are created. We ensure this by making sure
nodes are always associated with the owning BVH, and by having
pre-allocated indexes for invalid look-ups, i.e., handles will always
be able to point to a valid object.
With this change an additional ~15% performance gain was seen.

Change-Id: I879ea27042acaefb80f68d3685c83af551cde9dc
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>