Fix some issues with LodManager
The LodManager had a couple of issues that would cause the incorrect
level of detail to be used:
1) It was only listening to the Camera's local position for changes. If
the camera is attached to something else its local position would not
change but its scene position could. So listen to the scene position
instead.
2) If the scene does not update after the first frame, a LOD models
will be visible. We need to update the LOD/Distance calculations as
soon as the component is loaded so that only one LOD is shown on the
first frame.

Pick-to: 6.6 6.5
Change-Id: I942574f631109134a44187748ef5667e3ab542d2
Reviewed-by: Christian Strømme <christian.stromme@qt.io>