Release cached resource when the scene graph is invalidated
This amends 78ee5c95b48cb which introduced a call to cleanupResources()
which has serious performance implications for large sets of models due
to its bookkeeping. To avoid this we make sure we take the fast path and
drop all cached resources by calling releaseCachedResources() first.
This is the same approach as was done for QTBUG-114110 (890ec084da54)

Change-Id: I68565700200b0f6bbba1d2ac8ef03e43518feb15
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>