Ensure we don't reap textures registered by the extensions
Textures registered by extensions should not be automatically reaped
even when the "effective" usage count goes to zero. This is solved by
not reducing the usage count below 1 for textures as long as the
QQuick3DRenderExtension instance is alive and the registered texture is
not null. User can unregister a texture by registering a "null" texture
for the extension.

Change-Id: I38206188766ab1b6a59d70d246047472710c1558
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>