Update lightmapper log messages for model
Currently using debugObjectName which, if not set, just says
QQuick3DModel in the logs. It's now updated to just use the assigned key
instead, as this value is required to be unique for all models in the
scene that's part of the lightmap baking.

Fixes: QTBUG-118775
Pick-to: 6.5
Change-Id: Ie3933b076b34910310384d6dce5b13db06a4fdc6
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit d42f82f83df25fcf8cdff5923a2e069e0ca5691b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>