Fix lightmapper crash when castsShadow is false
If a model has a BakedLightmap that is enabled, but castsShadows is
false, the lightmapper crashes because of a missing check. Only models
with castsShadows set to true gets their geometry committed to the
RTCScene, which is when they get assigned a geomid. This change adds the
missing check.

Fixes: QTBUG-118773
Pick-to: 6.5
Change-Id: Ibe440d6a5068dca453053dc2eb22a0b3d51d8542
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit b65c17ed3136e555a92ba94a2bb0f0c8cb3805fe)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>