Enable docs from runtimerender directory
runtimerender/extensionapi is listed, so that everything
under it is scanned by qdoc, however for other semi-public
classes that live in other locations under runtimerender
we list the .h and .cpp files individually. This is good
for performance (no need to scan everything) and perhaps
helps keeping a better handle on what exactly is semi-public
and what isn't.

To avoid qdoc warnings (errors), the patch also adds some
documentation for some things. A number of arguments are just
added with \a tags but no proper text. Every single argument
has to be documented as per-qdoc rules, and writing proper
documentation is not trivial. Leaving this as a future
exercise.

Change-Id: Ie153f1a11864c3a301ed965dfb460d3f4ef6fdb5
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>