Per camera storage
Extends the per extension storage with per camera storage as well,
allowing to properly store data specific to the camera being used.
Note that this still has some limitation, e.g., it's expected that the
camera instance is unchanged within a frame.
Since the frustum culling is now done when building the sorted node lists,
it's applied more widely (when enabled), which means, e.g., the shadow map
will also have its nodes culled. This is a problem due to how the
shadow map code is implemented, as it's already doing it wrong, it something
that needs to be fixed separately.

Taks-number: QTBUG-116570
Change-Id: I9c77648ed82635f14f26ab8260c8b3e7127dd0f1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>