Long live Qt Quick 3D XR
Import the 6.7 branch of qtquick3dxr from git.qt.io
(nothing to do with Qt 6.7, that branch is just where
all the recent work is)

This only includes the Quick3D.Xr module itself, with
one example and a manual test. (adds the 'simple'
example as xr_simple and the desktop-VR hybrid test
app) There are no changes to the source code itself,
apart from cleaning up the sanity bot suggested things,
and adding some missing namespace macros, and fixing
some warnings found on Linux.

However, the build system (CMake stuff) is slightly
different, although the logic, incl. the required
env.vars, are exactly the same as before.
(OPENXR_ROOT, OCULUS_OPENXR_MOBILE_SDK)

Structurally it's all under xr/quick3dxr, with
quick3dxr-openxr being the controlling feature in the
build system, as it is all 100% tied to OpenXR. This
may evolve later on.

Building is currently tested on Windows and Linux (it
is up to the user to build and install
https://github.com/KhronosGroup/OpenXR-SDK
with the appropriate build type, and set OPENXR_ROOT)
and on Android using the Oculus OpenXR Mobile SDK (which
is also up to the user to install and set
OCULUS_OPENXR_MOBILE_SDK)

The CI does not currently build any of this code as there
are no OpenXR headers and loader in the CI images atm.

From now on applications should need no special macros
and CMake boilerplate whatsoever, apart from pulling in
Quick3DXr and using a custom manifest on Android.

Change-Id: Ib25da67bbe3fc35d5e32fc733cef8ea3857b8390
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>