Add warning about using the DefaultMaterial in new code
The DefaultMaterial type was only meant as compatibility material for
converting applications to QtQuick3D 5.15 from the old 3D Studio
projects, this is no longer relevant, but the unfortunate naming of the
DefaultMaterial stuck around, which now can easily lead users into
believing that the DefaultMaterial is what they should be defaulting to.
The DefaultMaterial will therefore be marked as deprecated since 6.4
and warn user to use either the SpecularGlossyMaterial or the
PrincipledMaterial in new code.

Pick-to: 6.6 6.5
Change-Id: I0b0c8a6acfbf76487365e49d76096a838a86280b
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
(cherry picked from commit 276fa6a01fed4c16866037fe5aece2cf5513ad5c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>