Add basic Quick 3D (XR) multiview plumbing
On the Quick 3D XR side, multiview is opt-in at the moment.
This is enabled by setting the QT_QUICK3D_XR_MULTIVIEW
env.var. to a non-zero value. If the rhi backend says
multiview is not supported, it will flip back to disabled
after printing some warnings. To be evaluated later how
this should work in the long run, e.g. make it opt-out
instead of opt-in, or if we want to only keep the multiview
path. For now, with development in mind, keeping it opt-in
is the sane approach.

NOTE: Currently multiview mode uses the first view's pose and
fov for both images, and is thus not suitable for viewing in
a real headset.

Fixes the Vulkan integration to request the instance version
as we do elsewhere, this is important for multiview since it
is a Vulkan 1.1+ feature. Also only enable the validation
layer if the window (QSG_RHI_DEBUG_LAYER) says so.

On the Quick 3D side, do some more viewCount plumbing, and
use the appropriate texture types where necessary (array,
not 2D, e.g. for the depth texture)
The consequences (in particular the ones for custom materials
and effects with user-provided shader code relying on the
depth or screen textures) are to be investigated separately.
Also, post-processing effects are not currently supported in
multiview mode; attempting to hit that code path in
QQuick3DSGDirectRenderer::prepare() and render()
will fail in unforeseen ways.

The shader support is incomplete. For built-in shaders we
now specify MULTIVIEW in CMakeLists, but adding the .mv2qsb
suffix in QSSGShaderCache::loadBuiltinForRhi is not yet
implemented since it requires plumbing through the view
count somehow. For generated shaders, we request HLSL 6.1
since that's mandatory for multiview with D3D12, and have
a commented line deep down in the baker that allows forcing
a view count of 2 and thus test multiview mode with
generated shaders. Uncommenting that line allows running
the xr_simple example, so that both the cubes and the
Qt Quick text show up as expected.

NB when running with QT_QUICK3D_XR_MULTIVIEW,
QT_QUICK3D_NO_SHADER_CACHE_LOAD=1 should be set (also
afterwards) to prevent reusing cached shaders (this
will go away later once the view count is in the
shader keys and everywhere else)

Task-number: QTBUG-119310
Task-number: QTBUG-114887
Change-Id: I80d9a839e37720fa074420342cea1a28cff2df81
Reviewed-by: Andy Nichols <andy.nichols@qt.io>