Make Quick3DXR build on macOS
While one cannot run anything, building and installing
the OpenXR SDK the same way as on Windows or Linux, and
then building Quick 3D with full OpenXR support is in
fact perfectly feasible. We have to make sure the OpenGL
path is not hit, and fix a few clang-reported format
string complaints, but that's that.

Attempting to run anything will either lead to failing
with
'The Qt Quick Scenegraph is not using a supported RHI mode'
as we clearly cannot do anything with Metal or OpenGL on this
platform,
or, in case Qt was built with Vulkan support via MoltenVK
and the application has requested Vulkan at run time,
failing due to not having an OpenXR runtime.

Change-Id: I2706e28aa7d1a3da546e883d32cf9cba1343e48b
Reviewed-by: Christian Strømme <christian.stromme@qt.io>