Robustify lightmapper blendLine function
QVector2D::operator/= asserts to avoid a division by zero, this is
called during the post process step in the blendLine() function. This
assert seems to be triggered in some cases, so far only on MSVC. This
was reproducible consistently when Model source was set to the primitive
types #Cone and #Cylinder. Changed from using the QVector2D overload to
doing the division manually, but only in the case where it's valid to do
so, to keep the current logic and output the same.

Fixes: QTBUG-120629
Pick-to: 6.6
Change-Id: Ic15c4f2a9523b83cb4bf256016aa78a49542c82b
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
(cherry picked from commit 58dc732e6e735be927f5eba65874a409bc1492fa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>