Make viewCount part of the shader key
For default/principled and custom materials the view count
will now flow automatically all the way down via the shader
key, and leads to setting the QShaderBaker's view count.
In the (vertex) generator the view count is not used yet,
but it is at least now queriable on that level too.

Also add something for the built-in shaders as well (a
viewCount argument that can be used to decide the filename),
but the story for these shaders is not quite clear yet,
and none of them is made multiview aware (using the QSHADER
ifdefs like in Qt Quick materials) for the time being.

Start using a common pattern for the todo-ish notes, so that
they can be found by searching for ### multiview

Change-Id: I0317764054cec481e94ce56fee9a9be16a96fa64
Reviewed-by: Andy Nichols <andy.nichols@qt.io>