Replicate Vulkan debug message behavior via XR_EXT_debug_utils
Route to qDebug, instead of letting the OpenXR runtime print its
messages via its own unspecified ways. Replicate the behavior
of QtGui's Vulkan integration effectively, which makes sense also
because the XR extension is a carbon copy of the VK_EXT_debug_utils.

Later this could be extended to collect the messages and offer
the list of strings to Qt Quick UIs as well.

Change-Id: I9d8244d14c742002c722290e16127de8b50bf7a8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>