Enable core validation if available and requested
Requesting is done with the same env.var. as for Quick, unfortunately
we cannot currently hook this up to Quick's facilities so query
the env.var. ourselves for now. (QSG_RHI_DEBUG_LAYER)

As usual we handle things gracefully, so requesting validation
with no api layer present is not an error and the layer is
filtered out before trying to create the instance.

The core validation layer is built from
https://github.com/KhronosGroup/OpenXR-SDK-Source and is installed
together with the headers, loader, and test apps by default. It
is worth noting that this repo is not the same as
https://github.com/KhronosGroup/OpenXR-SDK because the latter is
effectively a minimal, compact mirror of the former with just the
headers and loader. So people setting up their environment with the
e.g. on Windows will not have a validation layer available. Might
be worth keeping in mind for docs later on.

To make the validation layer available when building
OpenXR-SDK-Source from source, one must also set
the env.var. XR_API_LAYER_PATH=c:\openxr\bin\api_layers
(or bin\api_layers under whereever it was installed)
before running applications, otherwise the extra api layer
DLLs are not picked up by OpenXR.

Note that validation currently fails presumably due to
extension dependencies not provided currectly. To be
checked separately.

Change-Id: I66cf5377cfc60ebda9d9cd039ad5cc4607fcbc91
Reviewed-by: Andy Nichols <andy.nichols@qt.io>