Improve the quaternion compare in RotationData
Increase the resolution of our quaternion compare function.
This was mainly a problem observed in VR, as small movements
would cause very noticeable stuttering due to certain movements
being to small for the compare function to register those as a changes.
Since we where already at the measurable limit, for the values used,
the calculation of the dot product was bumped to use double precision
instead. There's potential for meeting dragons here, but RotationData
should be more precision that QQuaternion, even when doing component
wise compare...

Fixes: QTBUG-121414
Change-Id: I46d414d9010b868c119003b67b3cb34409cf046d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>