Enable particles for multiview
Also fixes line particles on D3D12 which was broken
independently of XR or multiview.

Also remove some unnecessary multiview todo lines both
in particles and elsewhere. We do not currently plan to
address the C++-side calculations, e.g. when sorting based
on distance from the camera: currently this uses the first
(left eye camera), and will continue to do so. Later we could
think of something (calculate something based on both cameras
positions and directions etc.), but this is clearly different
from multiview support in the shaders.

Change-Id: Ibeebbff363d4a79fbbb4ece4e09fed1cb80aa74b
Reviewed-by: Christian Strømme <christian.stromme@qt.io>