XR: Fix warning from an OpenXR debug statement
This just removes an unnecessary session=%p line from a qDebug statement
because it causes a warning on some platforms. Basically we are trying
to print a XrSession, but depending on the platform it will either be a
pointer or a unsigned long long. So rather than doing a cast or #ifdef
for this debug output, we will just remove this because it is
unnecessary.

Change-Id: I0c945f2ed8c6bf39833775ad12594d24c5cd7f93
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>