Clean up QSSGBuiltInRhiShaderCache
The function that actually loads stuff employs the common m_rhiShaders
cache. This is a waste of time, given the sole caller is
QSSGBuiltInRhiShaderCache, which stores every result and returns them
in subsequent calls without ever calling loadBuiltin again.
(historically this used to be different I suppose as there were direct
calls to loadBuiltin from elsewhere too; this is not the case anymore)

releaseCachedResources() was never implemented for the built-in
shaders. This is now corrected.

getBuiltinRhiShader() can now correctly drop the old version in case
the viewCount is different than before.

Remove unnecessary public and friend stuff. Also drop the legacy
6.0-era 'rhi' comments and names.

Change-Id: Ia52d4f5205f69167ba344ccc96c51c5883f7c0dc
Reviewed-by: Christian Strømme <christian.stromme@qt.io>