Rename XrActor to XrOrigin
This is necessary because XrActor is a very misleading. It may lead you
to beleive that the position of the XrActor is the position of the
person weaing the HMD, but that is not correct. It is instead an
anchored position in the scene that the reference point of all tracked
objects. If you were to move around width the HMD, the XrActor does not
move at all. If you were to teleport the HMD to move to another position
in world space, you do in fact move this reference point.

So a better name that reflects the true nature of this point is XrOrigin.
This becomes the basis for the coordinate space that all tracted items
live in (HMD, Controllers, Hands, other spatial anchors).

Further refinement of this in follow up patches.

Change-Id: Ic4b68b5f23074efd3df6062827e5ca86fb9ded92
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>