Introduce XrCamera to represent the Head Mounted Display (HMD)
All tracked objects in the scene must be direct children of an XrOrigin.
The XrCamera component represents the location of the HDM, or the head
of the person wearing the HMD. Logically that is an abstract camera, and
you can set a few properties on this object to affect the projection of
the camera (near and far clip planes). It is possible to use this
XrCamera object to get the location and rotation of the HMD in the scene
since it is tracked.

The "real" cameras for steroscopic rendering are now QOpenXREyeCamera
objects where are also direct children of the XrOrigin. These eye
cameras always exist and are an internal implimentation detail of
rendering the scene in VR/XR.

Change-Id: I2e4c7f92b83da88e727ec36b7cf2b04b2a05e207
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>