Workaround to allow NoCulling with D3D12 with multiview to function to a degree
The doublesided logic makes D3D12 fail pipeline creation with

ID3D12Device::CreateGraphicsPipelineState: Internal runtime error validating PS stage in a PSOs used with view instancing. [ STATE_CREATION ERROR #1105: CREATEGRAPHICSPIPELINESTATE_RUNTIME_INTERNAL_ERROR]

This is due to gl_FrontFacing, that becomes more or less this with HLSL
via SPIRV-Cross:

static bool gl_FrontFacing;
...
struct SPIRV_Cross_Input
{
    ...
    bool gl_FrontFacing : SV_IsFrontFace;
};

We do not know why this makes it choke when view instancing (==
multiview) is in use.

In Vulkan for example, there seems to be no problems with using
gl_FrontFacing with multiview.

As it is not clear what we can do here, as a temporary measure
make the front facing calculation a dummy for HLSL when in multiview
mode.

This is normally impossible to achieve, since the generated shaders
are compiled to SPIR-V and then transpiled into various languages.
One cannot differentiate the code based on a target (e.g. for HLSL).

That's where shadertools' per-target compilation mode comes in, which
allows us generating the following:

\#if QSHADER_HLSL && QSHADER_VIEW_COUNT >= 2
     const float qt_facing = 1.0;
\#else
     const float qt_facing = gl_FrontFacing ? 1.0 : -1.0;
\#endif

The downside is the increased time spent in bake(). (since now
there is a separate compilation to SPIR-V happening for each target
language and version). Therefore we only enable this if the material
says it's both for multiview and requires doublesided logic.

Task-number: QTBUG-122142
Change-Id: Id597b70fa0952a823f0ada934941eeeff19f1d4d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>