XR: Add support for getting and setting the reference space
The XrOrigin's relationship to the tracked items depends on the
reference space used by the XrView.  Previously we hardcoded local
reference space. This leads to the HMD being roughly at the same level
as the XrOrigin (XZ). When supported you now have the option to use
stage reference space, which will place the XrOrigin somewhere on the
ground and the HMD will reflect a +Y value above it. The local floor
reference space attempts to give you the best of both, where the origin
is reflects all three on reset.

Local Floor is a bit special because normally it is an extension, but
when it's not present it is possible to emulate if the support for stage
is available. If stage isn't available though, the best that can be done
is we can hardcode an everage height value as a guess.

Change-Id: I96a6c1aa055fd3c4f919cb9624725d66ed43e014
Reviewed-by: Christian Strømme <christian.stromme@qt.io>