Make fixed foveation user controllable
We start out with high, when supported. Print what we are
doing so it is obvious from the logs. The property can later
on be changed to something else, which updates the swapchain
with the new setting. (presumably effective from the next frame)

When not supported, setting the property has no effect.

Fixes: QTBUG-122293
Change-Id: I2e4d07acef71f0fcaab26f272ca6a8841c165536
Reviewed-by: Andy Nichols <andy.nichols@qt.io>