Add naive MSAA support
First part: this only supports MSAA through the XrSwapchain, i.e. if
the OpenXR runtime has implicit support for sample counts > 1.

The scene environment's antialiasingMode has so far been ignored.
Make it handle MSAA. (SSAA and ProgressiveAA are still ignored)

When it comes to antialiasingQuality, Medium is mapped to 2x MSAA,
whereas High and VeryHigh both map to 4x at the moment. (not sure if
there is a point in bothering with higher, given the performance
implications, and we anyway clamp to the view config's reported
maxSwapchainSampleCount which is likely at most 4 in practice)

This is the simple approach, relying on OpenXR: OpenXR can give out
multisample textures from the XrSwapchain and then we do not need to
do anything more, apart from rendering with appropriate
multisample-aware pipelines, resolving and such is up to the
runtime/compositor. (and Qt Quick already handles things like
auto-creating a multisample depth-stencil texture that is required
when the color buffer is multisample)

However, we destroy and recreate the XrSwapchain at least twice when
requesting MSAA since we cannot avoid setting things up with sample
count 1 first, and any QML properties from the SceneEnvironment get
processed only afterwards. In practice this seems painless, though.

The catch is that support varies. Due to the automatic clamping to
what OpenXR reports, requesting 2 or 4 (Medium or High) survives
always (with one exception, see below), but it then uses 1 (no
MSAA). Use the logs to confirm what is going on.

Air Link with Oculus OpenXR: Max reported sample count from the
runtime is 4. Choosing 2 has some visible effect, while 4 has a more
clear smoothing effect on the jagged edges of a rotating cube.

Steam Link with SteamVR: max reported sample count from the runtime is
1. So no MSAA here, but there is some sort of smoothing going on
implicitly in this system, since the look (cube's jagged edges) is
markedly different from e.g. Air Link / Oculus out of the box
(i.e. not quite jagged), although this may have something to do with
the video streaming used.

VirtualDesktop with SteamVR or VDXR: max reported sample count from
the runtime is 1.

Simulator: max reported sample count from the runtime is 4, but trying
to use 2 or 4 does not work with any 3D API. The rendered results are
either missing or are wrong (looks like the runtime does not correctly
treat the multisample texture as multisample?)

Android native: max reported sample count from the runtime is 1.

Thus we will still need to implement MSAA on our own (render into an
MSAA texture, resolve into the non-multisample OpenXR-provided
texture). This is also what the Oculus Mobile SDK's samples do.  That
is done in a separate patch.

Task-number: QTBUG-122288
Change-Id: I3c88407b6d3ed57e436f91599dd988672b3aac15
Reviewed-by: Andy Nichols <andy.nichols@qt.io>