Do MSAA on our own
Instead of passing in a > 1 sample count to the XrSwapchain, do MSAA
the usual way, resolving into the XrSwapchain-provided non-multisample
textures/texture array. The enablers all come from Qt Quick
(QQuickRenderTarget's fromXxxxTexture variants).

The antialiasingQuality is now treated identically to View3D
(Medium=2, High=4, VeryHigh=8).

Also must unify that _SRGB format mapping. Now there is another
reason: to ensure that with MSAA we do not end up with a different
from in the (multisample) color texture and the (non-multisample,
XR-provided) resolve texture.  That is an instant error with D3D11 for
example when issuing the ResolveSubresource().

Note that there is still a performance hazard, although it is nothing
new. Just like the depth-stencil buffer, the intermediate multisample
color buffer gets dropped and a new one is created on every
setRenderTarget() because the QQuickWindow only sees that a new
QQuickRenderTarget was given, there is no reuse. This is to be handled
in some future patch somehow.

Certain configurations may exhibit some unexpected results on the Meta
XR Simulator at least (e.g. Vulkan appears brighter with MSAA
enabled?).  These cases are to be checked separately, after testing on
real devices as well.

Fixes: QTBUG-122288
Change-Id: Ic157d135a5d1f4cdb4b66bde542ce74d8b9d34bb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>