Allow Item2D (Embedded 2D Quick) to report pick results
Previously we went through the effort to calculate a pick result for all
Item2Ds in the scene, specifically for handling automatic input support.
But it's possible that with generic raycasting that a user would want to
know if the ray intersected with an Item2D, and specifically which one.
This change adds support for doing that.

Also because this change adds an enum to the pick result class, it needs
to be exposed properly to QML as a named value type (since it's a
Q_GADGET). But that also means correcting all the incorrect
documentation.

PickResult is actually "pickResult" since this is a value type and must
start with a lower-case letter.

Change-Id: Ib0512c12d3691e6d840c52701b80d10cedb5812f
Reviewed-by: Christian Strømme <christian.stromme@qt.io>