xr: Use the new facilities to deal with the sRGB texture problem
For Vulkan and D3D12, we use the new viewFormat argument, where we
pass in the stripped (non-SRGB) format, while reporting the true
texture format (typically, with OpenXR, SRGB) as format.

What we want to get is:
- XR provides a rgba8-srgb texture
- pass it on without lying, but specify rgba8 as our "view" format
- on the QRhi side this means creating VkImageView (or the equivalents)
  with the view format when rendering into the texture and also when
  setting up the multisample resolve. (instead of just using the
  texture's VkFormat)

Until now this was not strictly required and we could just lie about
the format so that the VkImageViews (or equivalents) used our
fake (non-SRGB) format. This does not with with MSAA and Vulkan
(with some implementations at least), so stop lying, which causes
Qt Quick to create an SRGB texture for the intermediate multisample
color buffer as well. (and so rendering is rgba8-srgb viewed as
rgba8 in both the non-MSAA and MSAA case, while resolve is also
rgba8-srgb viewed as rgba8)

D3D11 does not support this since creating an RTV with a different type
would only be possible if the texture was created with a _TYPELESS
format (not _UNORM, not _SRGB, etc.). So instead, we continue to pass in
the non-sRGB format (e.g. RGBA8_UNORM) in place of the _SRGB one so that
QRhiTexture and everything else in QRhi thinks that it's UNORM, not
SRGB, and that's what the RTV is created with as well. This still
seems to work (not clear why), and avoids the unwanted sRGB conversions.

Use the new, unified fromXxxxTexture() overload in QQuickRenderTarget.
The previous approach with different fromXxxxTextureMultiView etc. is
dropped in favor of a single function that takes viewFormat, arraySize,
and flags.

Task-number: QTBUG-122288
Task-number: QTBUG-122614
Change-Id: I9c5c3ee92fa411b5299c00ee3cceb458c182f6f1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>